+++ src/main/java/com/dnsimple/request/DomainCheckPremiumPriceAction.java
@@ -1,5 +1,0 @@
-package com.dnsimple.request;
-
-public enum DomainCheckPremiumPriceAction {
-    REGISTRATION, RENEWAL, TRANSFER
-}